+++ packages/fluentui/react-northstar-prototypes/src/prototypes/VirtualizedStickyTree/useVirtualStickyTree.tsx
@@ -1,6 +1,28 @@
 import * as React from 'react';
-import { UseVirtualTreeOptions, UseVirtualTreeResult, useVirtualTree, TreeItemProps } from '@fluentui/react-northstar';
-import { VariableSizeListProps } from 'react-window';
+import {
+  UseVirtualTreeOptions,
+  UseVirtualTreeResult,
+  useVirtualTree,
+  TreeItemProps,
+  teamsTheme,
+  GetItemById,
+} from '@fluentui/react-northstar';
+import { ListChildComponentProps, VariableSizeListProps } from 'react-window';
+
+export interface InnerElementContextType {
+  getItemById: GetItemById;
+  stickyItemIds: string[];
+  stickyItemPusherHeights: number[];
+  stickyItemSize: number;
+  createTreeItem: (id: string, style: React.CSSProperties) => React.ReactElement<TreeItemProps> | null;
+}
+
+export interface VirtualItemData {
+  visibleItemIds: string[];
+  createTreeItem: (id: string, style: React.CSSProperties) => React.ReactElement<TreeItemProps> | null;
+}
+
+export const InnerElementContext = React.createContext<InnerElementContextType>({} as InnerElementContextType);
 
 export interface UseVirtualTreeStickyOptions extends UseVirtualTreeOptions, Pick<VariableSizeListProps, 'height'> {
   /** height of a non-sticky tree item */
@@ -227,3 +249,100 @@ export function useVirtualStickyTree(props: UseVirtualTreeStickyOptions): UseVir
     getItemOverrideProps,
   };
 }
+
+const getStickyItemStyle = (
+  indexAmoungStickyItems: number,
+  stickyItemNums: number,
+  stickyItemSize: number,
+): React.CSSProperties => ({
+  height: stickyItemSize,
+  zIndex: teamsTheme.siteVariables.zIndexes.overlay,
+  position: 'sticky',
+  top: indexAmoungStickyItems * stickyItemSize,
+  bottom: (stickyItemNums - indexAmoungStickyItems - 1) * stickyItemSize,
+  backgroundColor: teamsTheme.siteVariables.colorScheme.default.background3,
+});
+
+export const InnerElementType = React.forwardRef<HTMLDivElement, { style: React.CSSProperties }>((props, ref) => {
+  const { style, children, ...rest } = props;
+  const context = React.useContext(InnerElementContext);
+  const { stickyItemIds, stickyItemPusherHeights, stickyItemSize, getItemById, createTreeItem } = context;
+
+  const renderContent = React.useCallback(
+    (virtualItems: React.ReactElement<ListChildComponentProps>[]) => {
+      const result: Record<
+        string,
+        {
+          stickyItem: React.ReactElement; // the sticky item itself
+          pusher: React.ReactElement; // the div pusher with height being the same as all descendents of this sticky item
+          children: React.ReactElement[]; // all descendents of this sticky item
+        }
+      > = {};
+
+      stickyItemIds.forEach((id, index) => {
+        result[id] = {
+          stickyItem: createTreeItem(id, getStickyItemStyle(index, stickyItemIds.length, stickyItemSize)),
+          pusher: (
+            <div
+              key={`${id}-pusher`}
+              style={{ height: stickyItemPusherHeights[index], zIndex: -1 }}
+              role="presentation"
+            />
+          ),
+          children: [],
+        };
+      });
+
+      virtualItems.forEach(virtualItem => {
+        const virtualItemId = virtualItem.key as string; // our `getItemKey` makes virtual item's key the same as its corresponding tree item's id
+        // get the sticky id to which the current virtualItem belongs to
+        let parentId = getItemById(virtualItemId)?.parent;
+        let parentItem = getItemById(parentId);
+        while (parentItem && parentItem.level > 1) {
+          parentId = parentItem.parent;
+          parentItem = getItemById(parentId);
+        }
+        if (result[parentId] == null) {
+          return;
+        }
+        result[parentId].children.push(virtualItem);
+      });
+
+      const flattenedResult = [];
+      stickyItemIds.forEach(id => {
+        flattenedResult.push(result[id].stickyItem);
+        flattenedResult.push(result[id].pusher);
+        result[id].children.forEach(child => {
+          flattenedResult.push(child);
+        });
+      });
+
+      return flattenedResult;
+    },
+    [createTreeItem, getItemById, stickyItemIds, stickyItemPusherHeights, stickyItemSize],
+  );
+
+  return (
+    <div ref={ref} {...rest} style={style} role="none">
+      {renderContent(children as React.ReactElement<ListChildComponentProps>[])}
+    </div>
+  );
+});
+
+export const OuterElementType = React.forwardRef<HTMLDivElement>((props, ref) => (
+  <div
+    ref={ref}
+    {...props}
+    // OuterElement display scrollbar, it get focus even without any tabindex. Therefore we need to set -1 to not receive focus.
+    // tabindex="-1" causes a11y tree computes label from content. We need to set hacky way empty aria-label.
+    tabIndex={-1}
+    aria-label=" "
+  />
+));
+
+// memorize to avoid unnecessary re-renders, for example on scrolling
+// recommended approach by react-window: https://react-window.now.sh/#/api/FixedSizeList
+export const ItemWrapper = React.memo<ListChildComponentProps & { data: VirtualItemData }>(({ index, style, data }) => {
+  const { visibleItemIds, createTreeItem } = data;
+  return createTreeItem(visibleItemIds[index], style);
+});
